pyproject.toml
[project]
name = "nhmesh-producer"
version = "0.1.0"
description = "A resilient Meshtastic MQTT producer with automatic reconnection and connection health monitoring"
authors = [
    {name = "grleblanc",email = "greg@glsec.us"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "meshtastic @ git+https://github.com/meshtastic/python.git@master",
    "protobuf-to-dict (>=0.1.0,<0.2.0)",
    "ruff (>=0.11.7,<0.12.0)",
    "elasticsearch (>=8.0.0,<10.0.0)",
    "prettytable (>=3.16.0,<4.0.0)",
    "networkx (>=3.0.0,<4.0.0)",
    "pyvis (>=0.3.2,<1.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "folium (>=0.19.6,<0.20.0)",
    "flask (>=3.0.0,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

target-version = "py313"
indent-width = 2